# static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API UserHUB",
    "description": "Uma API de Gerenciamento de Usuários projetada para lidar com informações de usuários e departamentos, incluindo criação, recuperação, atualização e exclusão.",
    "version": "1.0.0",
    "contact": {
      "name": "Patrick Oliveira",
      "email": "patrickoliveiramoto@gmail.com",
      "url": "https://www.linkedin.com/in/patricksoares-dev/"
    }
  },
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
    "/api/auth/login": {
      "post": {
        "tags": ["Autenticação"],
        "summary": "Autenticar usuário",
        "description": "Autentica o usuário com email e senha, retornando um token JWT.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "description": "Email do usuário"
                  },
                  "senha": {
                    "type": "string",
                    "description": "Senha do usuário"
                  }
                },
                "required": ["email", "senha"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Autenticação bem-sucedida",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string",
                      "description": "Token JWT para autenticação"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Credenciais inválidas"
          }
        }
      }
    },
    "/api/auth/alterar_senha": {
      "post": {
        "tags": ["Autenticação"],
        "summary": "Alterar senha do usuário",
        "description": "Permite que um usuário autenticado altere sua senha.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nova_senha": {
                    "type": "string",
                    "description": "Nova senha do usuário"
                  }
                },
                "required": ["nova_senha"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Senha alterada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "response": {
                      "type": "string",
                      "example": "Senha alterada com sucesso"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos"
          },
          "403": {
            "description": "Token não fornecido ou inválido"
          }
        }
      }
    },
...
  }
}